requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-WTF>=1.2
python-telegram-bot>=20.7
google-generativeai>=0.5.2
